+++ fastai/__init__.py
@@ -8,7 +8,10 @@ from .layers import *
 from .metrics import *
 from .torch_core import *
 from .train import *
-from .gen_doc.nbdoc import doc
 from .datasets import *
 from .utils.collect_env import *
 from .version import __version__
+
+# Optional if jupyter is present
+try: from .gen_doc.nbdoc import doc
+except: pass